// src/app/api/mhdb/dataTypes/AmmoKind.ts
export enum AmmoKind {
  normal = "normal",
  pierce = "pierce",
  spread = "spread",
  slicing = "slicing",
  sticky = "sticky",
  cluster = "cluster",
  wyvern = "wyvern",
  poison = "poison",
  paralysis = "paralysis",
  sleep = "sleep",
  flaming = "flaming",
  water = "water",
  freeze = "freeze",
  thunder = "thunder",
  dragon = "dragon",
  recover = "recover",
  demon = "demon",
  armor = "armor",
  exhaust = "exhaust",
  tranq = "tranq",
}