// src/app/api/mhdb/dataTypes/Color.ts
export enum Color {
  white = "white",
  gray = "gray",
  rose = "rose",
  pink = "pink",
  red = "red",
  vermilion = "vermilion",
  orange = "orange",
  brown = "brown",
  ivory = "ivory",
  yellow = "yellow",
  lemon = "lemon",
  "sage-green" = "sage-green",
  "moss-green" = "moss-green",
  green = "green",
  emerald = "emerald",
  sky = "sky",
  blue = "blue",
  ultramarine = "ultramarine",
  "blue-purple" = "blue-purple",
  purple = "purple",
  "dark-purple" = "dark-purple",
}